Refactor : DDD 리팩토링에 따른 변경 파일 및 테스트코드 주석처리
