+++ orury-client/src/main/java/org/orury/client/auth/service/AuthService.java
@@ -1,6 +1,5 @@
 package org.orury.client.auth.service;
 
-import lombok.RequiredArgsConstructor;
 import org.orury.client.auth.jwt.JwtTokenProvider;
 import org.orury.common.error.code.AuthErrorCode;
 import org.orury.common.error.code.UserErrorCode;
@@ -8,12 +7,14 @@ import org.orury.common.error.exception.AuthException;
 import org.orury.common.error.exception.BusinessException;
 import org.orury.domain.auth.dto.JwtToken;
 import org.orury.domain.auth.dto.SignUpDto;
-import org.orury.domain.user.db.model.User;
-import org.orury.domain.user.db.repository.UserRepository;
-import org.orury.domain.user.dto.UserDto;
+import org.orury.domain.user.domain.dto.UserDto;
+import org.orury.domain.user.domain.entity.User;
+import org.orury.domain.user.infrastucture.UserRepository;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 @Service
