+++ orury-client/src/main/java/org/orury/client/auth/strategy/AppleLoginStrategy.java
@@ -1,23 +1,20 @@
 package org.orury.client.auth.strategy;
 
-import io.jsonwebtoken.Jwts;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.bouncycastle.asn1.pkcs.PrivateKeyInfo;
 import org.bouncycastle.openssl.jcajce.JcaPEMKeyConverter;
 import org.orury.client.auth.converter.message.AuthMessage;
 import org.orury.client.auth.converter.request.LoginRequest;
-import org.orury.client.auth.strategy.applefeign.AppleAuthClient;
 import org.orury.client.auth.jwt.JwtTokenProvider;
+import org.orury.client.auth.strategy.applefeign.AppleAuthClient;
 import org.orury.common.error.code.AuthErrorCode;
 import org.orury.common.error.exception.AuthException;
 import org.orury.common.util.TokenDecoder;
 import org.orury.domain.auth.dto.JwtToken;
 import org.orury.domain.auth.dto.LoginDto;
 import org.orury.domain.auth.dto.apple.AppleIdTokenPayload;
-import org.orury.domain.user.db.model.User;
-import org.orury.domain.user.db.repository.UserRepository;
-import org.orury.domain.user.dto.UserDto;
+import org.orury.domain.user.domain.dto.UserDto;
+import org.orury.domain.user.domain.entity.User;
+import org.orury.domain.user.infrastucture.UserRepository;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
@@ -29,6 +26,10 @@ import java.util.Base64;
 import java.util.Date;
 import java.util.Objects;
 import java.util.Optional;
+
+import io.jsonwebtoken.Jwts;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
 @Service
 @RequiredArgsConstructor
