+++ orury-client/src/main/java/org/orury/client/comment/controller/CommentController.java
@@ -11,13 +11,13 @@ import org.orury.client.comment.converter.response.CommentsWithCursorResponse;
 import org.orury.client.comment.service.CommentLikeService;
 import org.orury.client.comment.service.CommentService;
 import org.orury.client.post.service.PostService;
-import org.orury.client.user.service.UserService;
 import org.orury.domain.base.converter.ApiResponse;
 import org.orury.domain.comment.dto.CommentDto;
 import org.orury.domain.global.constants.NumberConstants;
+import org.orury.domain.user.domain.UserService;
+import org.orury.domain.user.domain.dto.UserDto;
+import org.orury.domain.user.domain.dto.UserPrincipal;
 import org.orury.domain.post.domain.dto.PostDto;
-import org.orury.domain.user.dto.UserDto;
-import org.orury.domain.user.dto.UserPrincipal;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
