+++ orury-client/src/main/java/org/orury/client/comment/controller/CommentLikeController.java
@@ -3,6 +3,7 @@ package org.orury.client.comment.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+
 import org.orury.client.comment.converter.message.CommentMessage;
 import org.orury.client.comment.service.CommentLikeService;
 import org.orury.client.comment.service.CommentService;
@@ -10,7 +11,7 @@ import org.orury.domain.base.converter.ApiResponse;
 import org.orury.domain.comment.db.model.CommentLike;
 import org.orury.domain.comment.db.model.CommentLikePK;
 import org.orury.domain.comment.dto.CommentLikeDto;
-import org.orury.domain.user.dto.UserPrincipal;
+import org.orury.domain.user.domain.dto.UserPrincipal;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
