+++ orury-client/src/main/java/org/orury/client/comment/converter/request/CommentCreateRequest.java
@@ -2,8 +2,8 @@ package org.orury.client.comment.converter.request;
 
 import org.orury.domain.comment.dto.CommentDto;
 import org.orury.domain.global.constants.NumberConstants;
+import org.orury.domain.user.domain.dto.UserDto;
 import org.orury.domain.post.domain.dto.PostDto;
-import org.orury.domain.user.dto.UserDto;
 
 public record CommentCreateRequest(
         String content,
