+++ orury-client/src/main/java/org/orury/client/post/interfaces/response/PostResponse.java
@@ -2,7 +2,7 @@ package org.orury.client.post.interfaces.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import org.orury.domain.post.domain.dto.PostDto;
-import org.orury.domain.user.dto.UserDto;
+import org.orury.domain.user.domain.dto.UserDto;
 
 import java.time.LocalDateTime;
 import java.util.List;
