+++ orury-client/src/main/java/org/orury/client/post/service/PostService.java
@@ -10,7 +10,7 @@ import org.orury.domain.global.domain.ImageUtils;
 import org.orury.domain.post.domain.entity.Post;
 import org.orury.domain.post.domain.dto.PostDto;
 import org.orury.domain.post.infrastructure.PostRepository;
-import org.orury.domain.user.dto.UserDto;
+import org.orury.domain.user.domain.dto.UserDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
