+++ orury-client/src/main/java/org/orury/client/review/converter/response/ReviewsResponse.java
@@ -1,9 +1,10 @@
 package org.orury.client.review.converter.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
+
 import org.orury.domain.global.constants.NumberConstants;
 import org.orury.domain.review.dto.ReviewDto;
-import org.orury.domain.user.dto.UserDto;
+import org.orury.domain.user.domain.dto.UserDto;
 
 import java.time.LocalDateTime;
 import java.util.List;
