+++ orury-client/src/test/java/org/orury/client/config/ControllerTest.java
@@ -18,11 +18,11 @@ import org.orury.client.review.controller.ReviewController;
 import org.orury.client.review.controller.ReviewReactionController;
 import org.orury.client.review.service.ReviewReactionService;
 import org.orury.client.review.service.ReviewService;
-import org.orury.client.user.controller.UserController;
-import org.orury.client.user.service.UserService;
+import org.orury.client.user.interfaces.UserController;
 import org.orury.common.config.SlackMessage;
 import org.orury.domain.gym.domain.GymService;
 import org.orury.domain.post.domain.PostLikeService;
+import org.orury.domain.user.domain.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
