+++ orury-domain/build.gradle
@@ -11,6 +11,7 @@ dependencies {
 
     //validation
     implementation 'org.springframework.boot:spring-boot-starter-validation'
+    implementation project(path: ':orury-common')
 
     //OAuth2, Security
     compileOnly 'org.springframework.boot:spring-boot-starter-oauth2-client'
