+++ orury-domain/src/main/java/org/orury/domain/comment/db/model/Comment.java
@@ -4,10 +4,10 @@ import jakarta.persistence.*;
 import lombok.*;
 import lombok.extern.slf4j.Slf4j;
 import org.orury.domain.base.db.AuditingField;
+import org.orury.domain.user.domain.entity.User;
 import org.orury.domain.global.constants.Constants;
 import org.orury.domain.global.constants.NumberConstants;
 import org.orury.domain.post.domain.entity.Post;
-import org.orury.domain.user.db.model.User;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import java.time.LocalDateTime;
