+++ orury-domain/src/main/java/org/orury/domain/comment/dto/CommentDto.java
@@ -1,8 +1,8 @@
 package org.orury.domain.comment.dto;
 
 import org.orury.domain.comment.db.model.Comment;
+import org.orury.domain.user.domain.dto.UserDto;
 import org.orury.domain.post.domain.dto.PostDto;
-import org.orury.domain.user.dto.UserDto;
 
 import java.time.LocalDateTime;
