+++ orury-domain/src/main/java/org/orury/domain/review/db/model/Review.java
@@ -9,7 +9,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.orury.domain.base.db.AuditingField;
 import org.orury.domain.global.listener.EntityImageConverter;
 import org.orury.domain.gym.domain.entity.Gym;
-import org.orury.domain.user.db.model.User;
+import org.orury.domain.user.domain.entity.User;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import java.time.LocalDateTime;
